Separate out invalid_state_transition_proof module in subspace-fraud-proof
Basically move the related code into a new file.
